Relocate messenger package
